backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
dnspython==2.8.0
email-validator==2.3.0
executing==2.2.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==309
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.48.0
tornado==6.4.2
traitlets==5.14.3
typer==0.19.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.36.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
